+++ spec/praxis/media_type_spec.rb
@@ -75,30 +75,9 @@ describe Praxis::MediaType do
     end
   end
 
-  context 'options' do
-    it 'does not have doc_visiblity set by default' do
-      expect(Address.options).to_not have_key(:doc_visiblity)
-    end
 
-    context 'with nodoc! called' do
-      subject(:media_type) do
-        Class.new(Praxis::MediaType) do
-          nodoc!
-          attributes do
-            attribute :name, String
-          end
-        end
-      end
-
-      it 'has the :doc_visibility set to :nodoc' do
-        expect(media_type.options[:doc_visibility]).to be(:nodoc)
-      end
-
-
-    end
-
-    context 'using blueprint caching' do
-      it 'has specs'
-    end
+  context 'using blueprint caching' do
+    it 'has specs'
   end
+
 end